# frontend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/liquidity/position/{nftMint}": {
      "get": {
        "description": "Возвращает детальную информацию о конкретной позиции",
        "operationId": "LiquidityBotController_getPositionInfo",
        "parameters": [
          {
            "name": "nftMint",
            "required": true,
            "in": "path",
            "description": "NFT Mint адрес позиции",
            "schema": {
              "example": "5HauEoJa6tcKBy1vofNAxnEDiMpC7H6V9L5TgLwFYNfd",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о позиции",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "position": {
                      "positionId": "5HauEoJa...",
                      "baseAmount": "21.526289",
                      "quoteAmount": "0.000000",
                      "priceRange": {
                        "lower": 154.63,
                        "upper": 161.81
                      },
                      "currentPrice": 162.27,
                      "profitability": 5.23
                    },
                    "pool": {
                      "poolId": "2Qdhep...",
                      "baseMint": "SOL",
                      "quoteMint": "USDC",
                      "currentPrice": 162.27
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Позиция не найдена"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Получить информацию о позиции",
        "tags": [
          "liquidity"
        ]
      }
    },
...
    "/transaction/position/{positionId}": {
      "get": {
        "description": "Возвращает все транзакции для конкретной позиции",
        "operationId": "TransactionController_getPositionHistory",
        "parameters": [
          {
            "name": "positionId",
            "required": true,
            "in": "path",
            "description": "ID позиции",
            "schema": {
              "example": "5HauEoJa6tcKBy1vofNAxnEDiMpC7H6V9L5TgLwFYNfd",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "История позиции",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "positionId": "5HauEoJa...",
                    "total": 2,
                    "transactions": [
                      {
                        "id": 1,
                        "type": "OPEN_POSITION",
                        "date": "2025-01-15T10:30:00Z"
                      },
                      {
                        "id": 2,
                        "type": "CLOSE_POSITION",
                        "date": "2025-01-16T14:20:00Z"
                      }
                    ]
                  }
                }
              }
            }
          },
          "404": {
            "description": "Позиция не найдена"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Получить историю позиции",
        "tags": [
          "transaction"
        ]
      }
    },
...
    "/analytics/position/{positionId}": {
      "get": {
        "operationId": "PositionAnalyticsController_getPositionAnalytics",
        "parameters": [
          {
            "name": "positionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PositionAnalytics"
        ]
      }
    },
    "/analytics/positions": {
      "get": {
        "operationId": "PositionAnalyticsController_getAllPositions",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PositionAnalytics"
        ]
      }
    },
...
    "/position-config": {
      "post": {
        "operationId": "PositionConfigController_upsertConfig",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConfigDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create or update position config",
        "tags": [
          "Position Config"
        ]
      },
      "get": {
        "operationId": "PositionConfigController_getAllConfigs",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all configs",
        "tags": [
          "Position Config"
        ]
      }
    },
    "/position-config/{poolId}": {
      "get": {
        "operationId": "PositionConfigController_getConfig",
        "parameters": [
          {
            "name": "poolId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get config for pool",
        "tags": [
          "Position Config"
        ]
      },
      "delete": {
        "operationId": "PositionConfigController_deleteConfig",
        "parameters": [
          {
            "name": "poolId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete config for pool",
        "tags": [
          "Position Config"
        ]
      }
    },
    "/position-config/{poolId}/deactivate": {
      "post": {
        "operationId": "PositionConfigController_deactivateConfig",
        "parameters": [
          {
            "name": "poolId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deactivate config for pool",
        "tags": [
          "Position Config"
        ]
      }
    },
    "/snapshots/position/{positionId}": {
      "get": {
        "operationId": "PositionSnapshotController_getPositionSnapshots",
        "parameters": [
          {
            "name": "positionId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "5ngdutp5g8cTbGFD4dV2hgGQPkPonp4gp32ZxPaMuiP",
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2025-01-01",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2025-01-31",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get snapshots for position",
        "tags": [
          "Position Snapshots"
        ]
      }
    },
...
    "/liquidity/optimal-range-volume/position/{positionId}": {
      "get": {
        "operationId": "OptimalRangeVolumeController_getForPosition",
        "parameters": [
          {
            "name": "positionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "period",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "OptimalRangeVolume"
        ]
      }
    },
...
  },
  "info": {
    "title": "Liquidity Bot API",
    "description": "API для управления ликвидностью на Raydium CLMM",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateConfigDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}